requirements.txt
# Core web framework
fastapi==0.115.0
uvicorn[standard]==0.30.6

# Slack
slack-bolt==1.20.0
slack-sdk==3.31.0

# LangChain
langchain==0.3.0
langchain-groq==0.2.0
langchain-core==0.3.0

# Database
asyncpg==0.29.0

# Utils
python-dotenv==1.0.1
httpx==0.27.2
pydantic==2.9.0